:art: Updated the folder structure
