+++ src/components/Form/Wrapper.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Heading from "../Header/Heading";
 import Input from "../Input/Input";
-import InputFile from "../InputFile/InputFile";
+import FileUpload from "../FileUpload/FileUpload";
 import FormWrapper from "../../styled/FormWrapper.style";
 
 const Wrapper = () => {
@@ -9,7 +9,7 @@ const Wrapper = () => {
     <div>
         <FormWrapper>
       <Heading />
-      <InputFile />
+      <FileUpload />
       <Input />
       </FormWrapper>
     </div>
