+++ src/components/Pages/Home/Navigation.style.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const Navigation = styled.div`
+height: 70px;
+background-color: #9cc3c9;
+display: flex;
+justify-content: flex-end;
+margin-bottom: 50px;
+`;
+
+export default Navigation;
