+++ src/components/Pages/Home/Paragraph.style.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const HomeParagraph = styled.p`
+font-size: 16px;
+color: black;
+`;
+
+
+export const HomeSuccessMsg = styled.p`
+font-size: 20px;
+font-weight: 500;
+color: green;
+`;
+
+export const HomeBold = styled.p`
+font-weight: 500;
+`;
