+++ src/components/Pages/Home/ProfilePicture.style.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const ProfilePicture = styled.img`
+height: 150px;
+width: 150px;
+`;
+
+export default ProfilePicture;
