+++ src/components/Pages/SignUp/SignUp.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import AppStyle from "../../styled/App.style";
-import Wrapper from "../Form/Wrapper";
-import Image from "../Image/Image";
+import AppStyle from "../../../styled/App.style";
+import Wrapper from "../../Form/Wrapper";
+import Image from "../../Image/Image";
 
 const SignUp = () => {
   return (
