+++ src/styled/WrapperStyle.style.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const WrapperStyle = styled.div`
+display: flex;
+justify-content: center;
+align-items: center;
+margin:10px;
+`;
+
+export default WrapperStyle;
